% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.Q.slug.R
\name{def.calc.Q.slug}
\alias{def.calc.Q.slug}
\title{Salt-based discharge calculations for a slug injection}
\usage{
def.calc.Q.slug(inputFile, slugMass = inputFile$slugTracerMass,
  slugVol = inputFile$carboyVolume, dataDir = paste0(getwd(),
  "/NEON_discharge-stream-saltbased.zip"), site = "all")
}
\arguments{
\item{inputFile}{Name of the data fram containing the information needed to calculate 
discharge from a slug tracer injection. If the headers are named: "slugMass" and 
"slugPourTime" the slugMass and slugPourTime paramaters don't need to be defined. 
Otherwise, the names of the columns need to be input for the function to work.}

\item{slugMass}{Mass of the tracer injectate [g]}

\item{slugVol}{Volume of the tracer injectate [L]}

\item{dataDir}{User identifies the directory that contains the zipped data}

\item{site}{User identifies the site(s), defaults to "all" [string]}
}
\value{
This function returns stream discharge [lps] appended as an additional column 
to the input data frame along with a quality flag (slugQF) where 1 means that a peak 
wasn't detected, 2 means the rising limb was not resolved, and 3 means the falling 
limb did not return to basline concentration.
}
\description{
This function calculates stream discharge from a slug salt tracer injection. 
This function will likely only work well for the NEON conductivity logger data.
}
\examples{
#sbdDataPlusSlugQ <- def.calc.Q.slug(inputFile = sbdFormatted)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.Q.inj.R for calculating stream discharge from a constant rate injection
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{discharge,}
\keyword{rivers,}
\keyword{salt-based,}
\keyword{slug}
\keyword{streams,}
\keyword{surface}
\keyword{tracer,}
\keyword{water,}
