% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.format.Q.R
\name{def.format.Q}
\alias{def.format.Q}
\title{Stream discharge formatting}
\usage{
def.format.Q(
  backgroundFieldCondData = NULL,
  backgroundDataSalt = NULL,
  fieldDataSite = NULL,
  plateauDataCond = NULL,
  plateauDataSalt = NULL,
  externalLabDataSalt = NULL
)
}
\arguments{
\item{backgroundFieldCondData}{Field data for conductivity timeseries, from 
sbd_backgroundFieldCondData table for NEON data [dataframe]}

\item{backgroundDataSalt}{Field data with background conductivity data, from 
sbd_backgroundFieldCondData table for NEON data [dataframe]}

\item{fieldDataSite}{Field data with experiment type information, from sbd_fieldData 
table for NEON data [dataframe]}

\item{plateauDataCond}{Field data with plateau conductivity data, from 
sbd_plateauMeasurementFieldData table for NEON data [dataframe]}

\item{plateauDataSalt}{Field data for salt sample data, from sbd_plateauSampleFieldData 
table for NEON data [dataframe]}

\item{externalLabDataSalt}{External lab data for injectate, background, and plateau, from 
sbd_externalLabDataSalt table for NEON data [dataframe]}
}
\value{
This function returns one data frame formatted for use with def.calc.Q.slug.R 
and def.calc.Q.constantRate.R to calculate stream discharge
}
\description{
This function formats data from the NEON Salt-based discharge (DP1.20193.001)
data product to calculate stream discharge from slug and constant-rate injections.
}
\examples{
#Using the example data in this package
#dataDirectory <- paste(path.package("neonStreamQ"),"inst\\extdata", sep = "\\")
#sbdFormatted <- def.format.Q(dataDir = dataDirectory)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.Q.slug.R and def.calc.Q.inj.R to calculate stream discharge
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{constant}
\keyword{discharge,}
\keyword{rate}
\keyword{rivers,}
\keyword{salt-based,}
\keyword{slug,}
\keyword{streams,}
\keyword{surface}
\keyword{tracer,}
\keyword{water,}
